+++ src/Traits/ComponentUtilities.php
@@ -2,6 +2,7 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Str;
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\ComponentConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\ComponentHelpers;
@@ -13,6 +14,7 @@ trait ComponentUtilities
 
     public array $table = [];
     public $theme = null;
+    protected Builder $builder;
     protected $model;
     protected $primaryKey;
     protected string $tableName = 'table';