+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 ### Added
 
 - Added table event listeners to sort/filter/clear from other components.
+- Added text filters.
 
 ## [2.3.0] - 2022-04-28