+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 - Added text filter.
 - Added $row as second parameter to BooleanColumn `setCallback()`.
 
+### Changed
+
+- Rework builder to fix passed parameters in `builder()` and `columns()` methods.
+- Fixed possible bug with bulk actions dropdown on Tailwind when bulk actions are hidden until a selection is made.
+
 ## [2.3.0] - 2022-04-28
 
 ### Added